#1387: Resolve conflicts with rebasing on upstream develop branch: moving new `TextCorpus` sampling method tests into `test_corpora`.
