+++ gensim/test/test_corpora.py
@@ -349,27 +349,46 @@ class TestTextCorpus(CorpusTestCase):
 
         return self.corpus_class(fpath)
 
-    def test_sample(self):
-        lines = ["document1", "document2"]
+    def test_sample_text(self):
+        lines = ["document%d" % i for i in range(10)]
         corpus = self.corpus_from_lines(lines)
         corpus.tokenizer = lambda text: text.split()
+        docs = [doc for doc in corpus.get_texts()]
 
         sample1 = list(corpus.sample_texts(1))
         self.assertEqual(len(sample1), 1)
-        document1 = sample1[0] == ["document1"]
-        document2 = sample1[0] == ["document2"]
-        self.assertTrue(document1 or document2)
+        self.assertIn(sample1[0], docs)
 
-        sample2 = list(corpus.sample_texts(2))
-        self.assertEqual(len(sample2), 2)
-        self.assertEqual(sample2[0], ["document1"])
-        self.assertEqual(sample2[1], ["document2"])
+        sample2 = list(corpus.sample_texts(len(lines)))
+        self.assertEqual(len(sample2), len(corpus))
+        for i in range(len(corpus)):
+            self.assertEqual(sample2[i], ["document%s" % i])
 
         with self.assertRaises(ValueError):
-            list(corpus.sample_texts(3))
+            list(corpus.sample_texts(len(corpus) + 1))
 
         with self.assertRaises(ValueError):
             list(corpus.sample_texts(-1))
+
+    def test_sample_text_length(self):
+        lines = ["document%d" % i for i in range(10)]
+        corpus = self.corpus_from_lines(lines)
+        corpus.tokenizer = lambda text: text.split()
+
+        sample1 = list(corpus.sample_texts(1, length=1))
+        self.assertEqual(sample1[0], ["document0"])
+
+        sample2 = list(corpus.sample_texts(2, length=2))
+        self.assertEqual(sample2[0], ["document0"])
+        self.assertEqual(sample2[1], ["document1"])
+
+    def test_sample_text_seed(self):
+        lines = ["document%d" % i for i in range(10)]
+        corpus = self.corpus_from_lines(lines)
+
+        sample1 = list(corpus.sample_texts(5, seed=42))
+        sample2 = list(corpus.sample_texts(5, seed=42))
+        self.assertEqual(sample1, sample2)
 
     def test_save(self):
         pass
